First TileMap ver completed
TileMap reading from file without handling errors (assuming that the file is perfect)
adjusting texture width, height, mapping textures to characters and making a map with corresponding characters in file.

The reading part is a complete disaster imo but I couldn't even create a switch because I can't put it in a switch so there is a lot of if else.

I am not sure how big of a deal this is since the code is only executed once