Created a First version of Particle Spawner
A simple particle spawner. Reading from file not implemented yet.
For the next commit I will make it so that the initial rotation of particles can be random and all the parameters of the particle spawner can be set and read from file.

Along with that I will re-make the reading with enums because currently it is a disaster